+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Change log
 1.2.10
 ------
 
+* 🎁 Append METAR information to clipboard file `clip.txt`
 * 💊 Limit variable wind bearing impact relative to wind force
 
 1.2.9